+++ ks_includes/screen_panel.py
@@ -10,28 +10,35 @@ logger = logging.getLogger("KlipperScreen.ScreenPanel")
 
 class ScreenPanel:
     title_spacing = 50
+    control = {}
 
     def __init__(self, screen, title, back=True):
         self._screen = screen
+        self._config = screen._config
+        self._files = screen.files
         self.lang = self._screen.lang
         self._printer = screen.printer
         self.labels = {}
+        self._gtk = screen.gtk
 
         self.layout = Gtk.Layout()
         self.layout.set_size(self._screen.width, self._screen.height)
 
+        button_scale = self._gtk.get_header_image_scale()
+        logger.debug("Button scale: %s" % button_scale)
         if back == True:
-            b = KlippyGtk.ButtonImage('back', None, None, 40, 40)
-            b.connect("clicked", self._screen._menu_go_back)
-            self.layout.put(b, 0, 0)
+            self.control['back'] = self._gtk.ButtonImage('back', None, None, button_scale[0], button_scale[1])
+            self.control['back'].connect("clicked", self._screen._menu_go_back)
+            self.layout.put(self.control['back'], 0, 0)
 
-            h = KlippyGtk.ButtonImage('home', None, None, 40, 40)
-            h.connect("clicked", self.menu_return, True)
-            self.layout.put(h, self._screen.width - 55, 0)
+            self.control['home'] = self._gtk.ButtonImage('home', None, None, button_scale[0], button_scale[1])
+            self.control['home'].connect("clicked", self.menu_return, True)
+            self.layout.put(self.control['home'], self._screen.width - round(
+                self._gtk.get_image_width() * button_scale[0] * 1.45), 0)
 
-        e = KlippyGtk.ButtonImage('emergency', None, None, 40, 40)
-        e.connect("clicked", self.emergency_stop)
-        self.layout.put(e, int(self._screen.width/4*3) - 20, 0)
+        self.control['estop'] = self._gtk.ButtonImage('emergency', None, None, button_scale[0], button_scale[1])
+        self.control['estop'].connect("clicked", self.emergency_stop)
+        self.layout.put(self.control['estop'], int(self._screen.width/4*3 - button_scale[0]/2), 0)
 
         self.title = Gtk.Label()
         self.title.set_size_request(self._screen.width, self.title_spacing)
@@ -56,6 +63,11 @@ class ScreenPanel:
     def get(self):
         return self.layout
 
+    def get_file_image(self, filename, width=1.6, height=1.6):
+        if not self._files.has_thumbnail(filename):
+            return None
+        return self._gtk.PixbufFromFile(self._files.get_thumbnail_location(filename), None, width, height)
+
     def home(self, widget):
         self._screen._ws.klippy.gcode_script(KlippyGcodes.HOME)
 
@@ -77,10 +89,13 @@ class ScreenPanel:
     def set_title(self, title):
         self.title.set_label(title)
 
+    def show_all(self):
+        self._screen.show_all()
+
     def update_image_text(self, label, text):
         if label in self.labels and 'l' in self.labels[label]:
             self.labels[label]['l'].set_text(text)
 
     def update_temp(self, dev, temp, target):
         if dev in self.labels:
-            self.labels[dev].set_label(KlippyGtk.formatTemperatureString(temp, target))
+            self.labels[dev].set_label(self._gtk.formatTemperatureString(temp, target))