+++ panels/bed_level.py
@@ -6,41 +6,63 @@ gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk, Gdk, GLib
 
 from ks_includes.KlippyGcodes import KlippyGcodes
-from ks_includes.KlippyGtk import KlippyGtk
 from ks_includes.screen_panel import ScreenPanel
 
 logger = logging.getLogger("KlipperScreen.BedLevelPanel")
 
+def create_panel(*args):
+    return BedLevelPanel(*args)
+
 class BedLevelPanel(ScreenPanel):
-    def initialize(self, menu):
+    x_offset = 0
+    y_offset = 0
+
+    def initialize(self, panel_name):
         _ = self.lang.gettext
+        self.panel_name = panel_name
         self.screws = None
-        grid = KlippyGtk.HomogeneousGrid()
+        grid = self._gtk.HomogeneousGrid()
         self.disabled_motors = False
 
         screws = []
-        if "bed_screws" in self._screen.printer.get_config_section_list():
-            bed_screws = self._screen.printer.get_config_section("bed_screws")
-            for item in bed_screws:
-                if re.match(r"^[0-9\.]+,[0-9\.]+$", bed_screws[item]):
-                    screws.append(bed_screws[item].split(","))
+        config_section_name = None
+        if "screws_tilt_adjust" in self._screen.printer.get_config_section_list():
+            config_section_name = "screws_tilt_adjust"
+        elif "bed_screws" in self._screen.printer.get_config_section_list():
+            config_section_name = "bed_screws"
+
+        if config_section_name != None:
+            config_section = self._screen.printer.get_config_section(config_section_name)
+            for item in config_section:
+                logger.debug("Screws section: %s" % config_section[item])
+                result = re.match(r"([0-9\.]+)\s*,\s*([0-9\.]+)", config_section[item])
+                if result:
+                    screws.append([
+                        round(float(result.group(1)),2),
+                        round(float(result.group(2)),2)
+                    ])
 
             screws = sorted(screws, key=lambda x: (float(x[1]), float(x[0])))
             logger.debug("Bed screw locations [x,y]: %s", screws)
-            if "bltouch" in self._screen.printer.get_config_section_list():
+            if ("bltouch" in self._screen.printer.get_config_section_list() and
+                    config_section_name == "screws_tilt_adjust"):
                 x_offset = 0
                 y_offset = 0
                 bltouch = self._screen.printer.get_config_section("bltouch")
                 if "x_offset" in bltouch:
-                    x_offset = float(bltouch['x_offset'])
+                    self.x_offset = float(bltouch['x_offset'])
                 if "y_offset" in bltouch:
-                    y_offset = float(bltouch['y_offset'])
+                    self.y_offset = float(bltouch['y_offset'])
                 new_screws = []
                 for screw in screws:
-                    new_screws.append([float(screw[0]) + x_offset, float(screw[1]) + y_offset])
+                    new_screws.append([
+                        round(float(screw[0]) + self.x_offset,2),
+                        round(float(screw[1]) + self.y_offset,2)
+                    ])
                 screws = new_screws
 
             self.screws = screws
+            logger.debug("Screws: %s" % screws)
 
         if len(screws) < 4:
             logger.debug("bed_screws not configured, calculating locations")
@@ -59,34 +81,41 @@ class BedLevelPanel(ScreenPanel):
             logger.debug("Configured screw locations [x,y]: %s", screws)
 
 
-        self.labels['tl'] = KlippyGtk.ButtonImage("bed-level-t-l")
-        self.labels['tl'].connect("clicked", self.go_to_position, self.screws[2])
-        self.labels['tr'] = KlippyGtk.ButtonImage("bed-level-t-r")
-        self.labels['tr'].connect("clicked", self.go_to_position, self.screws[3])
-        self.labels['bl'] = KlippyGtk.ButtonImage("bed-level-b-l")
-        self.labels['bl'].connect("clicked", self.go_to_position, self.screws[0])
-        self.labels['br'] = KlippyGtk.ButtonImage("bed-level-b-r")
-        self.labels['br'].connect("clicked", self.go_to_position, self.screws[1])
+        self.labels['bl'] = self._gtk.ButtonImage("bed-level-t-l", None, None, 3, 3)
+        self.labels['bl'].connect("clicked", self.go_to_position, self.screws[2])
+        self.labels['br'] = self._gtk.ButtonImage("bed-level-t-r", None, None, 3, 3)
+        self.labels['br'].connect("clicked", self.go_to_position, self.screws[3])
+        self.labels['fl'] = self._gtk.ButtonImage("bed-level-b-l", None, None, 3, 3)
+        self.labels['fl'].connect("clicked", self.go_to_position, self.screws[0])
+        self.labels['fr'] = self._gtk.ButtonImage("bed-level-b-r", None, None, 3, 3)
+        self.labels['fr'].connect("clicked", self.go_to_position, self.screws[1])
 
-        grid.attach(self.labels['tl'], 1, 0, 1, 1)
-        grid.attach(self.labels['tr'], 2, 0, 1, 1)
-        grid.attach(self.labels['bl'], 1, 1, 1, 1)
-        grid.attach(self.labels['br'], 2, 1, 1, 1)
+        grid.attach(self.labels['bl'], 1, 0, 1, 1)
+        grid.attach(self.labels['br'], 2, 0, 1, 1)
+        grid.attach(self.labels['fl'], 1, 1, 1, 1)
+        grid.attach(self.labels['fr'], 2, 1, 1, 1)
 
-        self.labels['home'] = KlippyGtk.ButtonImage("home",_("Home All"),"color2")
+        self.labels['home'] = self._gtk.ButtonImage("home",_("Home All"),"color2")
         self.labels['home'].connect("clicked", self.home)
 
-        self.labels['dm'] = KlippyGtk.ButtonImage("motor-off", _("Disable XY"), "color3")
+        self.labels['dm'] = self._gtk.ButtonImage("motor-off", _("Disable XY"), "color3")
         self.labels['dm'].connect("clicked", self.disable_motors)
 
         grid.attach(self.labels['home'], 0, 0, 1, 1)
         grid.attach(self.labels['dm'], 0, 1, 1, 1)
 
-        self.labels['estop'] = KlippyGtk.ButtonImage("decrease",_("Emergency Stop"),"color4")
-        self.labels['estop'].connect("clicked", self.emergency_stop)
-        grid.attach(self.labels['estop'], 3, 0, 1, 1)
+        if self._printer.config_section_exists("screws_tilt_adjust"):
+            self.labels['screws'] = self._gtk.ButtonImage("refresh",_("Screws Adjust"),"color4")
+            self.labels['screws'].connect("clicked", self.screws_tilt_calculate)
+            grid.attach(self.labels['screws'], 3, 0, 1, 1)
 
         self.content.add(grid)
+
+    def activate(self):
+        self.labels['bl'].set_label("")
+        self.labels['br'].set_label("")
+        self.labels['fl'].set_label("")
+        self.labels['fr'].set_label("")
 
     def go_to_position(self, widget, position):
         logger.debug("Going to position: %s", position)
@@ -110,5 +139,29 @@ class BedLevelPanel(ScreenPanel):
             "M18" # Disable motors
         )
 
-def create_panel(*args):
-    return BedLevelPanel(*args)
+    def process_update(self, action, data):
+        if action == "notify_gcode_response":
+            result = re.match(
+                "^// (.*) : X ([0-9\.]+), Y ([0-9\.]+), Z [0-9\.]+ : Adjust -> ([CW]+ [0-9:]+)",
+                data
+            )
+            if result:
+                screw_labels = ['fl','fr','br','bl']
+                x = int(float(result.group(2)) + self.x_offset)
+                y = int(float(result.group(3)) + self.y_offset)
+                logger.debug(data)
+                logger.debug("X: %s Y: %s" % (x,y))
+                for i in range(len(self.screws)):
+                    logger.debug(self.screws[i])
+                    if x == int(float(self.screws[i][0])) and y == int(float(self.screws[i][1])):
+                        break
+                self.labels[screw_labels[i]].set_label(result.group(4))
+                self.response_count += 1
+                if self.response_count >= len(self.screws)-1:
+                    self._screen.remove_subscription(self.panel_name)
+
+
+    def screws_tilt_calculate(self, widget):
+        self.response_count = 0
+        self._screen.add_subscription(self.panel_name)
+        self._screen._ws.klippy.gcode_script("SCREWS_TILT_CALCULATE")